[Tree widget]: Move caches under common folder
